dashboard/requirements.txt
streamlit
altair
psycopg2-binary
pandas
dotenv